Changed balance and amount to decimal in database

+++ java/src/main/java/org/ikigaidigital/adapter/in/web/dto/TimeDepositResponseDTO.java
@@ -1,5 +1,6 @@
 package org.ikigaidigital.adapter.in.web.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import org.ikigaidigital.domain.model.PlanType;
@@ -7,6 +8,6 @@ import org.ikigaidigital.domain.model.PlanType;
 import lombok.Builder;
 
 @Builder
-public record TimeDepositResponseDTO(Integer id, PlanType planType, Double balance, Integer days,
+public record TimeDepositResponseDTO(Integer id, PlanType planType, BigDecimal balance, Integer days,
 		List<WithdrawalsResponseDTO> withdrawals) {
 }
