+++ java/src/main/java/org/ikigaidigital/adapter/in/web/dto/WithdrawalsResponseDTO.java
@@ -1,9 +1,10 @@
 package org.ikigaidigital.adapter.in.web.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import lombok.Builder;
 
 @Builder
-public record WithdrawalsResponseDTO(Long id, Double amount, LocalDate date) {
+public record WithdrawalsResponseDTO(Integer id, BigDecimal amount, LocalDate date) {
 }
