+++ java/src/main/java/org/ikigaidigital/application/query/model/TimeDepositView.java
@@ -1,5 +1,6 @@
 package org.ikigaidigital.application.query.model;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import org.ikigaidigital.domain.model.PlanType;
@@ -7,6 +8,6 @@ import org.ikigaidigital.domain.model.PlanType;
 import lombok.Builder;
 
 @Builder
-public record TimeDepositView(Integer id, PlanType planType, Double balance, Integer days,
+public record TimeDepositView(Integer id, PlanType planType, BigDecimal balance, Integer days,
 		List<WithdrawalView> withdrawals) {
 }
