+++ java/src/main/java/org/ikigaidigital/application/query/model/WithdrawalView.java
@@ -1,9 +1,10 @@
 package org.ikigaidigital.application.query.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import lombok.Builder;
 
 @Builder
-public record WithdrawalView(Long id, Double amount, LocalDate date) {
+public record WithdrawalView(Integer id, BigDecimal amount, LocalDate date) {
 }
