+++ java/src/main/java/org/ikigaidigital/domain/service/InterestStrategy.java
@@ -1,5 +1,7 @@
 package org.ikigaidigital.domain.service;
 
+import java.math.BigDecimal;
+
 import org.ikigaidigital.domain.model.PlanType;
 import org.ikigaidigital.domain.model.TimeDeposit;
 
@@ -7,6 +9,6 @@ public interface InterestStrategy {
 
 	PlanType supports();
 
-	Double calculateInterest(TimeDeposit deposit);
+	BigDecimal calculateInterest(TimeDeposit deposit);
 
 }
